requirements.txt
pyautogui
Pillow
pytesseract
mistralai 
python-dotenv
requests
streamlit
SpeechRecognition 
pyaudio